refactor(httpapi): split event route modules
